KernelAbstractions 0.10 as extension
