+++ CUDACore/src/CUDACore.jl
@@ -17,7 +17,6 @@ using LLVM
 using LLVM.Interop
 using Core: LLVMPtr
 
-import KernelAbstractions
 import KernelInterface
 
 using Adapt: Adapt, adapt, WrappedArray
@@ -124,14 +123,10 @@ include("sorting.jl")
 include("complex.jl")
 include("library_types.jl")
 
-# KernelAbstractions
-include("CUDAKernelsOld.jl")
+# KernelInterface
+include("CUDAKernels.jl")
 import .CUDAKernels: CUDABackend
 export CUDABackend
-
-# KernelInterface - NOT PUBLIC. Use KernelInterface.get_backend on an CLArray to get the backend
-include("CUDAKernels.jl")
-import .CUDAInterface
 
 # StaticArrays is still a direct dependency, so directly include the extension
 include("../ext/StaticArraysExt.jl")
