+++ lib/cusparse/Project.toml
@@ -32,7 +32,7 @@ CUDA_Runtime_Discovery = "2"
 CUDA_Runtime_jll = "0.24"
 GPUArrays = "11.5.14"
 GPUToolbox = "3"
-KernelAbstractions = "0.9.38"
+KernelAbstractions = "0.10"
 LinearAlgebra = "1"
 SparseArrays = "1"
 SparseMatricesCSR = "0.6.9"
