+++ test/core/kernelinterface.jl
@@ -1,6 +1,6 @@
 import KernelInterface
-using CUDACore.CUDAInterface
+using CUDACore
 
 include(joinpath(dirname(pathof(KernelInterface)), "..", "test", "testsuite.jl"))
 
-Testsuite.testsuite(CUDAInterface.CUDABackend, "CUDACore", CUDACore, CuArray, CUDACore.CuDeviceArray)
+Testsuite.testsuite(CUDABackend, "CUDACore", CUDACore, CuArray, CUDACore.CuDeviceArray)
